% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_shiny.R
\name{alluvial_shiny}
\alias{alluvial_shiny}
\title{Tune and export an alluvial plot in Shiny}
\usage{
alluvial_shiny(x, run = TRUE, ...)
}
\arguments{
\item{x}{An \code{alluvial_model} object returned by \code{alluvial_prep_trace()} or \code{alluvial_prep_flow()}.}

\item{run}{Logical; if \code{TRUE}, run the app with \code{shiny::runApp()}. If \code{FALSE}, return the Shiny app object.}

\item{...}{Additional arguments passed to \code{shiny::runApp()} when \code{run = TRUE}.}
}
\value{
A Shiny app object when \code{run = FALSE}; otherwise the result of \code{shiny::runApp()}.
}
\description{
Tune and export an alluvial plot in Shiny
}
