% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_curve.R
\name{alluvial_curve}
\alias{alluvial_curve}
\title{Alluvial curve specification}
\usage{
alluvial_curve(from = -6, to = 6, length.out = 49)
}
\arguments{
\item{from}{integer specifying start point of sequence}

\item{to}{integer specifying end point of sequence}

\item{length.out}{positive integer specifying length of curve}
}
\value{
numeric vector
}
\description{
Alluvial curve specification
}
