% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model.R
\name{alluvial_model}
\alias{alluvial_model}
\title{Alluvial model specification}
\usage{
alluvial_model(x, pos)
}
\arguments{
\item{x}{data frame/tibble containing position information}

\item{pos}{character vector of column in df containing position information}
}
\value{
numeric vector of model specification
}
\description{
Alluvial model specification
}
