% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_plot.R
\name{alluvial_plot}
\alias{alluvial_plot}
\title{Title}
\usage{
alluvial_plot(
  x,
  bar_width = 1L,
  hpad = 0,
  col = "y_from",
  xlabs = NULL,
  labels = NULL,
  ggtitle = NULL,
  y_scale = "prop",
  bar_clrs = NULL,
  flow_clrs = NULL,
  bar_alpha = 1L,
  trace_alpha = 0.3,
  flow_alpha = 0.7,
  show.legend = TRUE,
  auto_theme = FALSE,
  remove_y_axis = FALSE,
  trace_lwd = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
