% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_plotly.R
\name{alluvial_plotly}
\alias{alluvial_plotly}
\title{Title}
\usage{
alluvial_plotly(
  x,
  highlight_col = "black",
  highlight_on = "plotly_hover",
  opacityDim = 0.2,
  dynamic = FALSE,
  tooltip = x$ID,
  bar_width = 1L,
  hpad = 0,
  col = "y_from",
  xlabs = NULL,
  labels = NULL,
  ggtitle = NULL,
  y_scale = "prop",
  bar_clrs = NULL,
  flow_clrs = NULL,
  bar_alpha = 1L,
  trace_alpha = 0.3,
  flow_alpha = 0.7,
  show.legend = TRUE,
  auto_theme = FALSE,
  remove_y_axis = FALSE,
  trace_lwd = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
