% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_prep_flow.R
\name{alluvial_prep_flow}
\alias{alluvial_prep_flow}
\title{Title}
\usage{
alluvial_prep_flow(
  data,
  steps,
  weights,
  fctr_order = steps,
  keep_vars = FALSE,
  curve = alluvial_curve(),
  res = 1L,
  model_fun = alluvial_model,
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\value{

}
\description{
Title
}
