% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_prep_trace.R
\name{alluvial_prep_trace}
\alias{alluvial_prep_trace}
\title{Title}
\usage{
alluvial_prep_trace(
  data,
  id,
  steps,
  fctr_order = steps,
  is.long = FALSE,
  values = NULL,
  keep_vars = FALSE,
  add_flows = FALSE,
  curve = alluvial_curve(),
  res = 1L,
  model_fun = alluvial_model,
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\value{

}
\description{
Title
}
